refactor: extract shared series aggregation logic into useSeriesAggregation composable (#38)
* Initial plan

* refactor: extract series aggregation into useSeriesAggregation composable


* fix(useSeriesAggregation): add explicit imports for computed and useI18n


* refactor: move SeriesItem type to app/types/composables.ts


---------
