+++ app/pages/index.vue
@@ -5,7 +5,7 @@ import {
   TvHero,
   TvLabel,
 } from '@todovue/tv-ui'
-import type { BlogPost, CardConfig, CardLabel } from '@/types/composables'
+import type { CardConfig, CardLabel } from '@/types/composables'
 
 const { t } = useI18n()
 
@@ -46,71 +46,7 @@ const latestPosts = computed<CardConfig[]>(() => {
   return allCards.slice(1, 5)
 })
 
-const normalizeSeriesKey = (value: unknown): string => typeof value === 'string' ? value.trim().toLowerCase() : ''
-const getDateValue = (value: BlogPost['date']): number => new Date(value ?? 0).getTime()
-
-type HomeSeries = {
-  slug: string
-  title: string
-  description: string
-  path: string
-  cover: string
-  coverAlt: string
-  chapters: number
-  latestDate: number
-  firstOrder: number
-}
-
-const homeSeries = computed<HomeSeries[]>(() => {
-  const seriesMap = new Map<string, HomeSeries>()
-
-  blogStore.blogPosts.value.forEach((post) => {
-    const seriesSlug = normalizeSeriesKey(post.series)
-    if (!seriesSlug) return
-
-    const existing = seriesMap.get(seriesSlug)
-    const postDate = getDateValue(post.date)
-    const postTitle = typeof post.seriesTitle === 'string' && post.seriesTitle.trim()
-      ? post.seriesTitle
-      : t('blogs.series.defaultTitle')
-    const postDescription = typeof post.seriesDescription === 'string' && post.seriesDescription.trim()
-      ? post.seriesDescription
-      : t('blogs.series.defaultDescription')
-    const postPath = `/series/${seriesSlug}/`
-
-    if (!existing) {
-      const initialOrder = typeof post.seriesOrder === 'number' ? post.seriesOrder : Number.MAX_SAFE_INTEGER
-      seriesMap.set(seriesSlug, {
-        slug: seriesSlug,
-        title: postTitle,
-        description: postDescription,
-        path: postPath,
-        cover: post.meta?.cover ?? '',
-        coverAlt: post.meta?.coverAlt ?? postTitle,
-        chapters: 1,
-        latestDate: postDate,
-        firstOrder: initialOrder
-      })
-      return
-    }
-
-    existing.chapters += 1
-    if (postDate > existing.latestDate) {
-      existing.latestDate = postDate
-    }
-
-    const currentOrder = typeof post.seriesOrder === 'number' ? post.seriesOrder : Number.MAX_SAFE_INTEGER
-    if (currentOrder <= existing.firstOrder && post.meta?.cover) {
-      existing.firstOrder = currentOrder
-      existing.cover = post.meta.cover
-      existing.coverAlt = post.meta?.coverAlt ?? postTitle
-    }
-  })
-
-  return Array.from(seriesMap.values())
-    .sort((first, second) => second.latestDate - first.latestDate)
-    .slice(0, 4)
-})
+const homeSeries = useSeriesAggregation(blogStore.blogPosts, 4)
 
 const seriesCards = computed<CardConfig[]>(() => homeSeries.value.map((series, index) => ({
   title: series.title,
